feat: Распределение по модулям
